Refactored NeumannConditions
